made the api work with doubles

+++ src/main/java/moze_intel/projecte/api/ProjectEAPI.java
@@ -18,11 +18,11 @@ public final class ProjectEAPI
 	 * Users can still modify inter-mod EMC registration via command/configuration file.<br>
 	 * Can be called during pre-init, init or post-init.
 	 */
-	public static void registerCustomEMC(ItemStack stack, int emcValue)
+	public static void registerCustomEMC(ItemStack stack, double emcValue)
 	{
 		NBTTagCompound nbt = new NBTTagCompound();
 		stack.writeToNBT(nbt);
-		nbt.setInteger("EMC", emcValue);
+		nbt.setDouble("EMC", emcValue);
 		
 		FMLInterModComms.sendMessage("ProjectE", "registeremc", nbt);
 	}
